refactor(projectsADM): move ProjectIRI value object + add optional handling
