+++ webapi/src/test/scala/org/knora/webapi/messages/admin/responder/valueObjects/BUILD.bazel
@@ -8,7 +8,7 @@ scala_test(
     size = "small",  # 60s
     srcs = [
         "GroupsValueObjectsADMSpec.scala",
-        "IriValueObjectsADMSpec.scala",
+        "ProjectsValueObjectsADMSpec.scala",
         "ListsValueObjectsADMSpec.scala",
         "UsersValueObjectsADMSpec.scala",
         "ValueObjectsADMSpec.scala",
